refactor: merge internal/history into internal/wiki
